@@ -2,18 +2,15 @@ import { contextBridge, ipcRenderer, IpcRendererEvent } from 'electron';
 
 export type Channels =
   | 'carplay-click'
-  | 'carplay-status-request'
-  | 'carplay-fps-request'
-  | 'carplay-reload-request'
   | 'carplay-quit-request'
-  | 'carplay-plugged'
-  | 'carplay-unplugged'
   | 'system-shutdown'
   | 'open-youtube'
   | 'open-netflix'
   | 'wifi-networks-request'
   | 'wifi-networks'
-  | 'wifi-connect';
+  | 'wifi-connect'
+  | 'wifi-current-request'
+  | 'wifi-current';
 
 contextBridge.exposeInMainWorld('electron', {
   ipcRenderer: {
